Updating

+++ CustomerFrequency/Model/rfm.py
@@ -1,8 +1,8 @@
 import pandas as pd
-
+import os
 # Load your transactional data from the database or CSV file
 
-def calculate_rfm_scores(transaction_data, output_file_path):
+def calculate_rfm_scores(transaction_data):
     """
     Calculate RFM scores for each customer based on transactional data.
 
@@ -34,15 +34,20 @@ def calculate_rfm_scores(transaction_data, output_file_path):
     # Sort the table based on RFM score
     rfm_table = rfm_table.sort_values(by='RFM_score', ascending=False)
 
-  
-
-
-    # Save the RFM table to a CSV file
-    rfm_table.to_csv(output_file_path)
-
     return rfm_table
 
+def save_rfm_csv(output_file_path,rfm_table):
+    # Save the RFM table to a CSV file
+    if os.path.exists(output_file_path):
+        os.remove(output_file_path)
+        rfm_table.to_csv(output_file_path)
+    else:
+        rfm_table.to_csv(output_file_path)
 
-transaction_data = pd.read_csv('Data/transactions_data.csv')
-output_file_path = ('Data/rfm_scores.csv')
-calculate_rfm_scores(transaction_data, output_file_path)
+
+
+# transaction_data = pd.read_csv('Data/transactions_data.csv')
+# output_file_path = ('Data/rfm_scores.csv')
+# table  = calculate_rfm_scores(transaction_data, output_file_path)
+# print(table)
+# print(table.dtypes)
